fix 64 bit warnings
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@2336 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCChasht.c
@@ -62,7 +62,7 @@ newHashTable (int size)
   if (!(htab->table = Safe_alloc ((size + 1) * sizeof (hashtItem *))))
     {
       fprintf (stderr, "out of virtual memory %s %d\n",
-	       __FILE__, (size + 1) * sizeof (hashtItem *));
+	       __FILE__, (size + 1) * (int) sizeof (hashtItem *));
       exit (1);
     }
   htab->minKey = htab->size = size;
